#1 refactor : 클래스 성격에 맞게 알맞은 패키지로 리팩토링 한다
